refactor: Promote the worker TLS material to the integration conftest
The CA and the two leaves it signs move out of test_identity_mtls into
the integration conftest as worker_certs, unchanged.

The shape of that material is what the mTLS tests actually assert on:
the worker leaf's only SAN is the logical identity, so it is
unverifiable by address and a handshake that succeeds can only have
succeeded through the identity override; the API leaf has no SAN at all
and clientAuth only, so a worker authenticating its client by chain
alone is exercised and the API's certificate cannot terminate a server
side. A second module needing the same material would otherwise
hand-copy eighty lines of X.509 minting, which is how two tests start
disagreeing about what a worker certificate is.

No behavioural change: the fixture body is identical and the existing
tests only take a new parameter name.
