+++ tests/integration/test_identity_mtls.py
@@ -5,11 +5,12 @@ certificates, because the thing under test *is* the TLS handshake — a
 mocked ``as_provider`` proves cfdb calls wool correctly but says nothing
 about whether the resulting channel actually comes up.
 
-The worker leaf minted here carries exactly one SAN, the logical
-identity — no ``localhost``, no ``127.0.0.1``. That is what makes the
-pair of tests meaningful: the certificate is unverifiable by address, so
-the dispatch that succeeds can only have succeeded through the identity
-override, and the one without an identity can only fail.
+The worker leaf (from the shared ``worker_certs`` fixture in
+``conftest``) carries exactly one SAN, the logical identity — no
+``localhost``, no ``127.0.0.1``. That is what makes the pair of tests
+meaningful: the certificate is unverifiable by address, so the dispatch
+that succeeds can only have succeeded through the identity override,
+and the one without an identity can only fail.
 
 The two ends hold separate leaves signed by a shared CA, as they do in
 deployment, which takes a little arranging: a spawning pool passes its
@@ -24,16 +25,11 @@ case where a mistake stays invisible.
 from __future__ import annotations
 
 import asyncio
-import datetime
 import logging
 import re
 
 import pytest
 import wool
-from cryptography import x509
-from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.x509.oid import ExtendedKeyUsageOID, NameOID
 
 from cfdb.workflows.credentials import DEFAULT_TLS_IDENTITY, build_worker_credentials
 
@@ -64,116 +60,7 @@ _SHUTDOWN_TIMEOUT_S = 15.0
 _FAILED_SHUTDOWN_TIMEOUT_S = 5.0
 
 
-def _generate_key() -> rsa.RSAPrivateKey:
-    # 2048 rather than the 4096 the shipped script uses: these certs live
-    # for one test, and key generation dominates its runtime.
-    return rsa.generate_private_key(public_exponent=65537, key_size=2048)
-
-
-def _common_name(value: str) -> x509.Name:
-    return x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, value)])
-
-
-def _sign(subject, subject_key, issuer, issuer_key, *, sans=None, eku=None, ca=False):
-    """Mint a certificate, self-signed when issuer and subject coincide."""
-    now = datetime.datetime.now(datetime.timezone.utc)
-    builder = (
-        x509.CertificateBuilder()
-        .subject_name(subject)
-        .issuer_name(issuer)
-        .public_key(subject_key.public_key())
-        .serial_number(x509.random_serial_number())
-        .not_valid_before(now - datetime.timedelta(minutes=5))
-        .not_valid_after(now + datetime.timedelta(days=1))
-        .add_extension(x509.BasicConstraints(ca=ca, path_length=None), critical=True)
-    )
-    if sans is not None:
-        builder = builder.add_extension(
-            x509.SubjectAlternativeName(sans), critical=False
-        )
-    if eku is not None:
-        builder = builder.add_extension(x509.ExtendedKeyUsage(eku), critical=False)
-    return builder.sign(issuer_key, hashes.SHA256())
-
-
-def _write(path, data: bytes) -> str:
-    path.write_bytes(data)
-    return str(path)
-
-
-def _write_leaf(tmp_path, stem, ca_name, ca_key, *, common_name, sans, eku):
-    """Mint a CA-signed leaf and return its ``(cert_path, key_path)``."""
-    key = _generate_key()
-    cert = _sign(_common_name(common_name), key, ca_name, ca_key, sans=sans, eku=eku)
-    return (
-        _write(
-            tmp_path / f"{stem}-cert.pem",
-            cert.public_bytes(serialization.Encoding.PEM),
-        ),
-        _write(
-            tmp_path / f"{stem}-key.pem",
-            key.private_bytes(
-                encoding=serialization.Encoding.PEM,
-                format=serialization.PrivateFormat.TraditionalOpenSSL,
-                encryption_algorithm=serialization.NoEncryption(),
-            ),
-        ),
-    )
-
-
-@pytest.fixture
-def certs(tmp_path):
-    """Mint a CA plus the separate worker and API leaves it signs.
-
-    Returns ``(ca_path, worker_cert, worker_key, api_cert, api_key)``,
-    mirroring the deployed arrangement where each process holds its own
-    leaf and only the CA is shared.
-
-    The worker leaf's sole SAN is the logical identity — no
-    ``localhost``, no ``127.0.0.1`` — so it is unverifiable by address.
-    The API leaf gets none at all, which is the point: a worker
-    authenticates its client by chain, with no name check.
-
-    The extended key usages mirror what ``certs/generate-worker-certs.sh``
-    now mints, so these tests fail if a ``clientAuth``-only API leaf turns
-    out not to work as a client. Withholding ``serverAuth`` from the API
-    is what stops its certificate doubling as a worker's.
-    """
-    ca_key = _generate_key()
-    ca_name = _common_name("cfdb-worker-ca")
-    ca_cert = _sign(ca_name, ca_key, ca_name, ca_key, ca=True)
-
-    worker_cert, worker_key = _write_leaf(
-        tmp_path,
-        "worker",
-        ca_name,
-        ca_key,
-        common_name="cfdb-worker",
-        sans=[x509.DNSName(DEFAULT_TLS_IDENTITY)],
-        # serverAuth to terminate dispatch; clientAuth because wool's
-        # graceful-stop RPC dials the worker's own subprocess.
-        eku=[ExtendedKeyUsageOID.SERVER_AUTH, ExtendedKeyUsageOID.CLIENT_AUTH],
-    )
-    api_cert, api_key = _write_leaf(
-        tmp_path,
-        "api",
-        ca_name,
-        ca_key,
-        common_name="cfdb-api",
-        sans=None,
-        eku=[ExtendedKeyUsageOID.CLIENT_AUTH],
-    )
-
-    return (
-        _write(tmp_path / "ca.pem", ca_cert.public_bytes(serialization.Encoding.PEM)),
-        worker_cert,
-        worker_key,
-        api_cert,
-        api_key,
-    )
-
-
-def _worker_factory(certs):
+def _worker_factory(worker_certs):
     """Build a worker factory that keeps the worker's own credentials.
 
     A spawning pool hands its single ``credentials=`` to both the proxy
@@ -192,7 +79,7 @@ def _worker_factory(certs):
     losing in-flight work — visible here as the pool warning that it
     "stopped waiting for 1 worker(s) that did not stop gracefully".
     """
-    ca, worker_cert, worker_key, _, _ = certs
+    ca, worker_cert, worker_key, _, _ = worker_certs
 
     def factory(*tags, credentials=None, host=None):
         return wool.LocalWorker(
@@ -206,12 +93,12 @@ def _worker_factory(certs):
     return factory
 
 
-def _pool(certs, *, identity, shutdown_timeout=_SHUTDOWN_TIMEOUT_S):
+def _pool(worker_certs, *, identity, shutdown_timeout=_SHUTDOWN_TIMEOUT_S):
     """Build a one-worker pool over the minted material."""
-    ca, _, _, api_cert, api_key = certs
+    ca, _, _, api_cert, api_key = worker_certs
     return wool.WorkerPool(
         spawn=1,
-        worker=_worker_factory(certs),
+        worker=_worker_factory(worker_certs),
         credentials=build_worker_credentials(
             ca, api_cert, api_key, identity=identity
         ),
@@ -234,7 +121,7 @@ def _pool(certs, *, identity, shutdown_timeout=_SHUTDOWN_TIMEOUT_S):
 class TestIdentityMutualTls:
     @pytest.mark.asyncio
     async def test_dispatch_should_succeed_when_identity_matches_the_certificate(
-        self, certs
+        self, worker_certs
     ):
         """Test that an identity makes an address-unverifiable cert usable.
 
@@ -249,7 +136,7 @@ class TestIdentityMutualTls:
             the address the pool dialed.
         """
         # Arrange
-        pool = _pool(certs, identity=DEFAULT_TLS_IDENTITY)
+        pool = _pool(worker_certs, identity=DEFAULT_TLS_IDENTITY)
 
         # Act
         async with pool:
@@ -262,7 +149,7 @@ class TestIdentityMutualTls:
 
     @pytest.mark.asyncio
     async def test_dispatch_should_fail_when_identity_is_not_configured(
-        self, certs, caplog
+        self, worker_certs, caplog
     ):
         """Test that the same certificate is unusable without an identity.
 
@@ -279,7 +166,7 @@ class TestIdentityMutualTls:
         """
         # Arrange
         pool = _pool(
-            certs, identity=None, shutdown_timeout=_FAILED_SHUTDOWN_TIMEOUT_S
+            worker_certs, identity=None, shutdown_timeout=_FAILED_SHUTDOWN_TIMEOUT_S
         )
 
         # Act & assert
